+++ test/altering_test.rb
@@ -15,4 +15,7 @@ class RailwayInsertTest < Minitest::Spec
   it { pipe.>(A, name: :A).>(C, prepend: true, name:  "C").inspect.must_equal %{[>C,>A]} }
   it { pipe.>(A, name: :A).>(C, replace: :A, name: :C).inspect.must_equal %{[>C]} }
   it { pipe.>(A, name: :A).add(nil, :A, delete: :A).inspect.must_equal %{[]} }
+  # last insert operation wins.
+  it { pipe.>(A, name: :A).>(B, name: :B).>(C, name: :C, after: :A, before: :A).inspect.must_equal %{[>C,>A,>B]} }
+  it { pipe.>(A, name: :A).>(B, name: :B).>(C, name: :C, before: :A, after: :A).inspect.must_equal %{[>A,>C,>B]} }
 end